+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release notes
 
+## 1.0.0
+
+Add support for other GitHub organizations then `grrr-amsterdam`.
+
+- Add `GRRR_WORDPRESS_PUBLISH_OWNER` and `GRRR_WORDPRESS_PUBLISH_REPOSITORY` constants
+- Remove hard coded grrr-amsterdam GitHub owner connection
+- Remove deprecated config vars starting with `GITHUB_DEPLOY_*`
+
 ## 0.3.0
 
 - Add scheduled publish support